# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScoutOps Internal API
  version: 0.1.0
  description: P00 foundation, P01 identity/tenancy/RBAC, and P02 scoped UI preference contracts. Organization-visible endpoints require server-side authorization.
servers:
  - url: /api/v1
...
paths:
...
  /auth/register:
    post:
      summary: Register a local email account and queue verification delivery
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LocalAccountRegistration' } } }
      responses:
        '201': { description: Pending-verification account without action or session token, content: { application/json: { schema: { $ref: '#/components/schemas/LocalAccountEnvelope' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /auth/email-verification/confirm:
    post:
      summary: Consume a single-use email verification token
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ActionTokenConfirm' } } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
  /auth/login:
    post:
      summary: Authenticate a verified local account and set an opaque HttpOnly session cookie
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LocalLoginRequest' } } } }
      responses:
        '200':
          description: Authenticated for the configured 30-day session lifetime; raw session token exists only in Set-Cookie
          headers: { Set-Cookie: { schema: { type: string }, description: '__Host-scoutops_session with Secure, HttpOnly, SameSite=Strict, Path=/ and Max-Age=2592000' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginEnvelope' } } }
        '202':
          description: Password accepted; TOTP or one-time recovery code is still required
          headers: { Set-Cookie: { schema: { type: string }, description: '__Host-scoutops_mfa_challenge; Secure; HttpOnly; SameSite=Strict; short lived' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MfaChallengeEnvelope' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '423': { $ref: '#/components/responses/Error' }
  /auth/mfa/totp/verify:
    post:
      summary: Complete a short-lived MFA login challenge and create a session
      security: [{ HttpOnlyMfaChallenge: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaCodeRequest' } } } }
      responses:
        '200':
          description: Challenge consumed once; challenge cookie cleared and session cookie set
          headers: { Set-Cookie: { schema: { type: array, items: { type: string } }, description: '30-day session cookie with Max-Age=2592000 plus expired MFA challenge cookie' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginEnvelope' } } }
        '401': { $ref: '#/components/responses/Error' }
        '423': { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      summary: Revoke the current local session and clear browser state
      security: [{ HttpOnlySession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '204': { description: Session revoked and cookie cleared }
        '401': { $ref: '#/components/responses/Error' }
  /auth/password-reset/request:
    post:
      summary: Queue a password reset if the account exists; response is enumeration-safe
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } } }
      responses:
        '202': { description: Always accepted for a syntactically valid email, content: { application/json: { schema: { $ref: '#/components/schemas/AcceptedEnvelope' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /auth/password-reset/confirm:
    post:
      summary: Consume a single-use reset token, change password and revoke sessions
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetConfirm' } } } }
      responses:
        '204': { description: Password changed; all sessions revoked }
        '400': { $ref: '#/components/responses/Error' }
  /me/password:
    post:
      summary: Change own password after current-password verification and revoke all sessions
      security: [{ HttpOnlySession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordChangeRequest' } } } }
      responses:
        '204': { description: Password changed and session cookie cleared }
        '401': { $ref: '#/components/responses/Error' }
  /me/security-setup:
    get:
      summary: Read the authenticated seed administrator's forced security setup state
      description: A restricted seed session may call only password, MFA, security-setup and logout endpoints until both steps complete.
      security: [{ HttpOnlySession: [] }]
      responses:
        '200': { description: Security setup posture without secrets, content: { application/json: { schema: { $ref: '#/components/schemas/SecuritySetupEnvelope' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /me/sessions:
    get:
      summary: List only the authenticated user's device sessions
      security: [{ HttpOnlySession: [] }]
      responses:
        '200': { description: Own sessions without token hashes, content: { application/json: { schema: { $ref: '#/components/schemas/SessionListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /me/mfa:
    get:
      summary: Read the authenticated user's MFA status and non-enabled enterprise identity capabilities
      security: [{ HttpOnlySession: [] }]
      responses:
        '200': { description: Own MFA status without encrypted secret or recovery hashes, content: { application/json: { schema: { $ref: '#/components/schemas/MfaStatusEnvelope' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /me/mfa/totp/enrollment:
    post:
      summary: Reauthenticate and create a pending encrypted TOTP factor
      security: [{ HttpOnlySession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaEnrollmentRequest' } } } }
      responses:
        '201': { description: Secret is displayed only for this enrollment response and must never be logged, content: { application/json: { schema: { $ref: '#/components/schemas/MfaEnrollmentEnvelope' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /me/mfa/totp/confirm:
    post:
      summary: Verify the pending factor and return one-time recovery codes once
      security: [{ HttpOnlySession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaCodeRequest' } } } }
      responses:
        '200': { description: TOTP enabled and recovery codes shown once, content: { application/json: { schema: { $ref: '#/components/schemas/MfaEnrollmentConfirmEnvelope' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /me/mfa/totp:
    delete:
      summary: Reauthenticate with password and current second factor, disable TOTP and revoke every session
      security: [{ HttpOnlySession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaDisableRequest' } } } }
      responses:
        '204': { description: TOTP disabled; all sessions and browser cookie revoked }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /me/sessions/{sessionId}:
    delete:
      summary: Revoke one session owned by the authenticated user
      security: [{ HttpOnlySession: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '204': { description: Owned session revoked }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    HttpOnlySession:
      type: apiKey
      in: cookie
      name: __Host-scoutops_session
    HttpOnlyMfaChallenge:
      type: apiKey
      in: cookie
      name: __Host-scoutops_mfa_challenge
    OpenApiClient:
      type: http
      scheme: bearer
      bearerFormat: scoutops-api-client
    ReleaseProbeHmac:
      type: apiKey
      in: header
      name: X-Release-Probe-Signature
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 255 }
...
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Error:
      description: Stable sanitized error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    SecuritySetupStatus:
      type: object
      additionalProperties: false
      required: [required, must_change_password, must_enroll_mfa, completed_at]
      properties:
        required: { type: boolean }
        must_change_password: { type: boolean }
        must_enroll_mfa: { type: boolean }
        completed_at: { type: string, format: date-time, nullable: true }
    SecuritySetupEnvelope:
      allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/SecuritySetupStatus' } } }]
...
    LocalAccountRegistration:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 1024, writeOnly: true }
    LocalLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, maxLength: 1024, writeOnly: true }
    ActionTokenConfirm:
      type: object
      additionalProperties: false
      required: [token]
      properties: { token: { type: string, minLength: 20, maxLength: 512, writeOnly: true } }
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties: { email: { type: string, format: email, maxLength: 254 } }
    PasswordResetConfirm:
      type: object
      additionalProperties: false
      required: [token, new_password]
      properties:
        token: { type: string, minLength: 20, maxLength: 512, writeOnly: true }
        new_password: { type: string, minLength: 8, maxLength: 1024, writeOnly: true }
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password: { type: string, maxLength: 1024, writeOnly: true }
        new_password: { type: string, minLength: 8, maxLength: 1024, writeOnly: true }
    LocalAccountSummary:
      type: object
      required: [id, email, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        status: { type: string, enum: [pending_verification, active, locked, disabled] }
    SessionSummary:
      type: object
      required: [id, status, device_label, expires_at, last_seen_at, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [active, revoked, expired] }
        device_label: { type: string }
        expires_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    MfaCodeRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties: { code: { type: string, minLength: 6, maxLength: 32, writeOnly: true } }
    MfaEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [current_password]
      properties: { current_password: { type: string, maxLength: 1024, writeOnly: true } }
    MfaDisableRequest:
      type: object
      additionalProperties: false
      required: [current_password, code]
      properties:
        current_password: { type: string, maxLength: 1024, writeOnly: true }
        code: { type: string, minLength: 6, maxLength: 32, writeOnly: true }
    MfaChallengeResponse:
      type: object
      required: [mfa_required, expires_at]
      properties:
        mfa_required: { type: boolean, enum: [true] }
        expires_at: { type: string, format: date-time }
    MfaEnrollmentStart:
      type: object
      required: [factor_id, secret, otpauth_uri]
      properties:
        factor_id: { type: string, format: uuid }
        secret: { type: string, description: Displayed once; never logged or persisted by the browser }
        otpauth_uri: { type: string, description: RFC 6238 authenticator enrollment URI displayed once }
    IdentityAdapterCapability:
      type: object
      required: [protocol, status, activation]
      properties:
        protocol: { type: string, enum: [oidc, saml2, scim2] }
        status: { type: string, enum: [adapter_ready, reserved_disabled] }
        activation: { type: string, enum: [requires_approved_provider_and_tenant_mapping] }
    MfaStatus:
      type: object
      required: [totp_enabled, factor_id, confirmed_at, identity_adapters]
      properties:
        totp_enabled: { type: boolean }
        factor_id: { type: string, format: uuid, nullable: true }
        confirmed_at: { type: string, format: date-time, nullable: true }
        identity_adapters: { type: array, items: { $ref: '#/components/schemas/IdentityAdapterCapability' } }
    LocalAccountEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { $ref: '#/components/schemas/LocalAccountSummary' } } }
    LoginEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - type: object
          properties:
            data:
              type: object
              required: [user, session]
              properties:
                user: { $ref: '#/components/schemas/LocalAccountSummary' }
                session: { $ref: '#/components/schemas/SessionSummary' }
                security_setup: { $ref: '#/components/schemas/SecuritySetupStatus' }
    MfaChallengeEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { $ref: '#/components/schemas/MfaChallengeResponse' } } }
    MfaEnrollmentEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { $ref: '#/components/schemas/MfaEnrollmentStart' } } }
    MfaEnrollmentConfirmEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - type: object
          properties:
            data:
              type: object
              required: [status, recovery_codes]
              properties:
                status: { type: string, enum: [enabled] }
                recovery_codes: { type: array, minItems: 4, items: { type: string } }
    MfaStatusEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { $ref: '#/components/schemas/MfaStatus' } } }
    AcceptedEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { type: object, required: [accepted], properties: { accepted: { type: boolean, enum: [true] } } } } }
    SessionListEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/SessionSummary' } } } }
...
    SuccessEnvelope:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data: {}
        request_id: { type: string }
        trace_id: { type: string }
    ErrorEnvelope:
      type: object
      required: [error, request_id, trace_id]
      properties:
        error:
          type: object
          required: [code, message, action_hint]
          properties:
            code: { type: string }
            message: { type: string }
            action_hint: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
...